feat: achieve 100% test pass rate and fix core logic bugs
- Resovled MagicMock leaks and pollution in unit tests
- Fixed ForeignKeyViolations by implementing robust teardowns
- Merged duplicate ghost dungeon cleanup logic in DungeonService
- Corrected stat scaling and base health calculation in UserService
- Standardized item effect keys (Nitro/Turbo)
- Added cleanup and rebuild scripts for test database
- Fixed crafting test dependencies and resource drops

+++ models/user.py
@@ -96,4 +96,4 @@ class Utente(Base):
 class Admin(Base):
     __tablename__ = "admin"
     id = Column(Integer, primary_key=True)
-    id_telegram = Column('id_telegram',Integer)
+    id_telegram = Column('id_telegram', BigInteger)
